Huge spawn weapon refactor
Let every gamemode set their own default weapon (weapon in hand on spawn)
And all the available weapons a tee spawns with.

Then also add support for sv_spawn_weapons which supports laser/grenade
but might later also accept custom weapon combos.

Setting sv_spawn_weapons only has an effect in zCatch so if it is set on
other gamemodes the user gets a warning in the console.
Based on this config the gamemode (for now only zCatch) still has to
derive the default weapon and set all the weapons.
To avoid duplicating code if there are multiple weapon combos and this
will be used by fng, tfng, idm, itdm and more there are two helper
methods to do so.
GetDefaultWeaponBasedOnSpawnWeapons() and SetSpawnWeapons()