stats: track amount of shots and hits
This also implicitly tracks misses (shots - hits)

Shots that hit the owner or team mates are not counted but their hits are.
This means that the amount of hits can be higher than the amount of shots.
This is a feature not a bug. Only projectile that go into the void are
counted as misses. Self or team boosts are never counted as shot so if
they do an additional kill it improves the accuracy stats a lot.

Because there are no accounts stats like accuracy or K/D ratio
can not be taken serious anyways because they can easily be destroyed
by fakers.