Overhaul console design: replace background, improve contrast
Replace the dotted background texture of the console with a less noticeable noise texture.

Replace the large striped bar at the bottom of the console with a small bar with a slightly darker color than the respective console background. Remove the shadow above the striped bar.

Improve the spacing between the console backlog and input with the additional free space from the replaced striped bar. The total console height is not changed.

Improve the contrast of the console completion options by slightly increasing the saturation/brightness of the text color and by coloring the selection rectangle black instead of using an orange color which was very similar to the highlight color. Also highlight the matching string of the selected completion option instead of only highlighting it for all other options.