Minor refactoring of server announcement loading
Move `reload_announcement` command handling from `CGameContext` to `CServer`. This is consistent with the conchain for `sv_announcement_filename` also being registered in `CServer`. Furthermore, this allows making the `ReadAnnouncementsFile` function non-`virtual` as it does not need to be exposed to the gameserver anymore.

Remove unnecessary filename argument of `ReadAnnouncementsFile` function, which is always set to the config variable.

Log number of loaded announcements on success. Use error log level on error.