Add parameter `AllowTruncation` to `CPacker::AddString` function
Previously, the `CPacker::AddString` function would truncate strings if the `int Limit` parameter or the packet buffer size were exceeded. When the packer buffer size is exceeded the packer would also enter the error state.

The parameter `bool AllowTruncation` is added to control whether the string is allowed to be truncated when using the `Limit` parameter. If truncation is not allowed then the packer will instead enter the error state and the truncated string will not be packed. The parameter `AllowTruncation` has the default value `true` to preserve the behavior of existing function calls, where the parameter is used to truncate strings.

The behavior of the `AddString` function when it enters the error state is changed so the packer position is not advanced to the end of the truncated string. This makes it consistent with the behavior of the other packer functions which either pack the entire given argument or do not change the packer position.

This allows using the `Limit` parameter to pack a message with a specific maximum size (i.e. the maximum payload size) without depending on the protocol specifics in the packer implementation. In particular, this is intended for cases where as many strings as possible are packed into a message, in which case the strings should not be truncated.