Improve entities/menu background map loading
Fix loading screen rendering without progress bar for the entities background map during the initial loading.

Remove the `bool RenderLoadingBar` parameter of the `CMenus::RenderLoading` function. The progress bar is now rendered as long as the loading total is set. Add the `CMenus::FinishLoading` function to reset the loading total at the end of the loading process.

Remove the `bool RenderMenuBackgroundMap` parameter of the `CMenus::RenderLoading` function. The `RenderLoading` function will now check internally if the menu background map can be rendered.