Remove obsolete and broken `MakeVersion` function for map items
The `MakeVersion` function dates back to at least 72c06a258940696093f255fb1061beb58e1cdd0b but it seems obsolete and broken in its current state. The function call `MakeVersion(1, *pTilemapItem)` currently returns the version number `65628` which does not conform to the version scheme for other map items and is not in the range of expected versions for `CMapItemLayerTilemap` items (versions 1-4). Because this functions derives the version based on the size of the `CMapItemLayerTilemap` item, it likely does not work correctly anymore, as the size of the map item has been changed after the `MakeVersion` function was created. Furthermore, additional handling for `CMapItemLayerTilemap` items of this version would also be missing on the server-side and for ingame client rendering.