Fix damage indicators being stuck after connecting dummy
Do not reset the local time after the dummy is connected, as the local time is supposed to run continuously while on the same server. Resetting the local time caused damage indicators to become stuck, as the `s_LastLocalTime` value is in the future after resetting the local time. This also affected other effects like pickup location and bullet rotation, where this was not noticeable.

Closes #4967.