Use separate packer with larger buffer only for Teehistorian
Revert buffer size for regular `CPacker` to 2 KiB to avoid stack overflows on systems with only 1 MiB of stack memory (e.g. Windows with MSVC and Android).

Add separate `CTeehistorianPacker` with a larger buffer to be used only for Teehistorian chunks.

To avoid duplicate code the existing `CPacker` is renamed to `CAbstractPacker` and the buffer and its size are supplied by the concrete subclasses `CPacker` and `CTeehistorianPacker`.