Fix Teehistorian server crash with MSVC due to uninitialized memory
The variable `CServer::CClient::m_Authed` and the function `IServer::GetAuthedState` may only be used when the client slot is not empty, as the authed state is uninitialized before a client slot has been used for the first time. When compiling the server with MSVC the variable is not zero-initialized like with other compilers currently. This was causing the server compiled with MSVC to crash on start when Teehistorian is enabled, as empty clients were incorrectly considered as authenticated for recording the initial auth state to the Teehistorian file. For these uninitialized clients, the `IServer::GetAuthName` function was returning `nullptr` which was then causing the `CPacker::AddString` function to crash when trying to pack the `nullptr` string.

Add check to only record the initial auth state for non-empty players.

Add assertions to prevent incorrect usage of the `IServer::GetAuthedState` and `IServer::GetAuthName` functions.

Replace the separate `IServer::ClientAuthed` function with the `IServer::GetAuthedState` function.