Fix menus skin list initially being empty, refactor skin refreshing
The skin list in the menus was initially empty because `s_SkinLastRefreshTime` is initialized to the current refresh time, which prevents the list from being updated unless the user already has favorite skins, as adding these also cause the list to be updated.

The static `s_SkinLastRefreshTime` variable is replaced with a member variable `m_SkinListNeedsUpdate` and an empty optional is used as initial value to always refresh the list.

The separate `m_SkinListNeedsUpdate` and `m_SkinFavoritesChanged` variables are removed and instead the `m_SkinListLastRefreshTime` variable is used to handle all updates of the skin list.

Separate refresh time variables `m_SkinList7LastRefreshTime` and `m_SkinPartsList7LastRefreshTime` are added for the 0.7 skin menu to avoid interference with 0.6 skins.