Support running local server on Android
Compile the server as a separate library for Android same as the client.

Use a foreground service to run the native `main` function of the server using Java Native Interface. Directly using JNI avoids the use of SDL as a wrapper for the server. The server service must run in a different process because it needs to be terminated to correctly restart the server. Otherwise it crashes because global variables are not initialized again when it is restarted. This also prevents simply running the server in another thread of the client process. A toast error message is shown on non-zero return values of server's `main` function.

A foreground service is used so the server can keep running while the client activity is in the background. Add detailed description why the service uses `android:foregroundServiceType="specialUse"` in the manifest as this should convince the reviewer that the use case is valid. None of the other foreground service types cover hosting local game servers or related use cases. The required permissions for using a foreground service with special use case are added as well as the permission to post notifications.

Showing a notification is necessary for foreground services to stay alive and it also allows showing the server status and controlling the server. In particular, an action to directly stop the server is shown. Another action to run commands based on user input in the notification is provided, which allows setting the initial rcon password.

The server will also be stopped automatically if the client is quit or if the server notification is deleted (which is possible on newer Android versions that do not allow ongoing notifications anymore).

Rename `NativeMain.java` to `ClientActivity.java` to differentiate the file better from the server class.