Make client/graphics warnings thread-safe, improve log format
Add mutex for client/graphics warnings. Return an optional copy of the current warning and remove it from the vector immediately instead of returning a pointer to it and removing it later.

Use atomic for `m_WarnPngliteIncompatibleImages` as this variable is also used in multiple threads.

Do not use `CLock` as the thread-safety annotations would otherwise have to be added to a lot of client/graphics functions.

Use `client` instead of the warning title as the log system, as the warning title is usually too long.

Closes #9354.